requirements.txt
fastapi
uvicorn
asyncpg
redis
pydantic
pydantic-settings
PyJWT